+++ src/compas_viewer/scene/lineobject.py
@@ -1,12 +1,7 @@
-from typing import List
-from typing import Optional
-from typing import Tuple
-
-from compas.colors import Color
 from compas.geometry import Line
-from compas.geometry import Point
 from compas.scene import GeometryObject
 
+from .sceneobject import DataType
 from .sceneobject import ViewerSceneObject
 
 
@@ -16,7 +11,7 @@ class LineObject(ViewerSceneObject, GeometryObject):
     def __init__(self, line: Line, **kwargs):
         super(LineObject, self).__init__(geometry=line, **kwargs)
 
-    def _read_points_data(self) -> Optional[Tuple[List[Point], List[Color], List[List[int]]]]:
+    def _read_points_data(self) -> DataType:
         positions = [self.geometry.start, self.geometry.end]
         colors = [
             self.pointscolor.get(0, self.pointscolor["_default"]),  # type: ignore
@@ -25,7 +20,7 @@ class LineObject(ViewerSceneObject, GeometryObject):
         elements = [[0], [1]]
         return positions, colors, elements
 
-    def _read_lines_data(self) -> Optional[Tuple[List[Point], List[Color], List[List[int]]]]:
+    def _read_lines_data(self) -> DataType:
         positions = [self.geometry.start, self.geometry.end]
         colors = [
             self.pointscolor.get(0, self.pointscolor["_default"]),  # type: ignore